+++ profiles.py
@@ -1,0 +1,32 @@
+profiles = {
+    "webdev": {
+        "profile" : "Web Development",
+        "description" : "Frontend and Backend Development",
+        "tools" : [
+        "git",
+        "node",
+        "npm",
+        "vscode"
+    ]},
+
+    "python": {
+        "profile" : "Python",
+        "description" : "Python Developmental Environment.",
+        "tools" : [
+        "python",
+        "pip",
+        "git"
+    ]
+
+    },
+
+    "ai": {
+        "profile" : "AI",
+        "description" : "AI Developmental Environment.",
+        "tools" : [
+        "python",
+        "pip",
+        "git",
+        "jupyter"
+    ]}
+}